+++ addon/nodes/index.ts
@@ -4,6 +4,9 @@ export { hard_break } from './hard-break';
 export { horizontal_rule } from './horizontal-rule';
 export { invisible_rdfa } from './invisible-rdfa';
 export { paragraph } from './paragraph';
-export { paragraphWithConfig } from './paragraphWithConfig';
+export {
+  paragraphWithConfig,
+  DERIVED_FROM_PARAGRAPH,
+} from './paragraphWithConfig';
 export { repaired_block } from './repaired-block';
 export { text } from './text';